requirements.txt
streamlit
langchain
langchain-groq
PyPDF2
python-dotenv
tiktoken
